+++ src/app/shared/PersonName.svelte
@@ -18,6 +18,7 @@
   import {themeColors} from "src/partials/state"
   import Popover from "src/partials/Popover.svelte"
   import Anchor from "src/partials/Anchor.svelte"
+  import WotScore from "src/partials/WotScore.svelte"
   import {displayPubkey} from "src/domain"
   import {userFollows, deriveProfileDisplay, session, maxWot, getWotScore} from "src/engine"
 
@@ -27,11 +28,7 @@
   const wotScore = getWotScore($session?.pubkey, pubkey)
   const npubDisplay = displayPubkey(pubkey)
   const profileDisplay = deriveProfileDisplay(pubkey)
-
-  $: superMaxWot = $maxWot * 1.5
-  $: dashOffset = 100 - (Math.max(superMaxWot / 20, wotScore) / superMaxWot) * 100
-  $: style = `transform: rotate(${dashOffset * 1.8 - 50}deg)`
-  $: stroke = $themeColors[$following || pubkey === $session?.pubkey ? "accent" : "neutral-200"]
+  const accent = $following || pubkey === $session?.pubkey
 </script>
 
 <div class={cx("flex gap-1", $$props.class)}>
@@ -44,22 +41,9 @@
   {#if $session}
     <div class="flex gap-1 font-normal">
       <Popover triggerType="mouseenter">
-        <span
-          slot="trigger"
-          class="relative flex h-10 w-10 items-center justify-center whitespace-nowrap px-4 text-xs">
-          <svg height="32" width="32" class="absolute">
-            <circle class="wot-background" cx="16" cy="16" r="15" />
-            <circle
-              cx="16"
-              cy="16"
-              r="15"
-              class="wot-highlight"
-              stroke-dashoffset={dashOffset}
-              {style}
-              {stroke} />
-          </svg>
-          {wotScore}
-        </span>
+        <div slot="trigger">
+          <WotScore score={wotScore} max={$maxWot} {accent} />
+        </div>
         <Anchor modal slot="tooltip" class="flex items-center gap-1" href="/help/web-of-trust">
           <i class="fa fa-info-circle" />
           WoT Score